# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "version": "1.0.0",
        "title": "GitHub Issues to Markdown",
        "description": "Translates GitHub issues list to Markdown format",
        "x-ibm-application-id": "gh-to-md",
        "x-ibm-skill-type": "imported"
    },
    "security": [
        {
            "basicAuth": []
        }
    ],
    "servers": [
        {
            "url": "https://application-a6.15v354wgljnn.us-east.codeengine.appdomain.cloud"
        }
    ],
    "paths": {
        "/markdown": {
            "post": {
                "summary": "Create a Markdown file from issues list",
                "description": "Create a Markdown file from issues list",
                "operationId": "markdown",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/issue"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "fileContent": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Bad request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/markdown_single": {
            "post": {
                "summary": "Create a Markdown file from issue",
                "description": "Create a Markdown file from issue",
                "operationId": "markdown",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/issue"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "fileContent": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Bad request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "author-association": {
                "title": "author_association",
                "type": "string",
                "example": "OWNER",
                "description": "How the author is associated with the repository.",
                "enum": [
                    "COLLABORATOR",
                    "CONTRIBUTOR",
                    "FIRST_TIMER",
                    "FIRST_TIME_CONTRIBUTOR",
                    "MANNEQUIN",
                    "MEMBER",
                    "NONE",
                    "OWNER"
                ]
            },
...
            "reaction-rollup": {
                "title": "Reaction Rollup",
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "total_count": {
                        "type": "integer"
                    },
                    "+1": {
                        "type": "integer"
                    },
                    "-1": {
                        "type": "integer"
                    },
                    "laugh": {
                        "type": "integer"
                    },
                    "confused": {
                        "type": "integer"
                    },
                    "heart": {
                        "type": "integer"
                    },
                    "hooray": {
                        "type": "integer"
                    },
                    "eyes": {
                        "type": "integer"
                    },
                    "rocket": {
                        "type": "integer"
                    }
                },
                "required": [
                    "url",
                    "total_count",
                    "+1",
                    "-1",
                    "laugh",
                    "confused",
                    "heart",
                    "hooray",
                    "eyes",
                    "rocket"
                ]
            },
            "nullable-simple-user": {
                "title": "Simple User",
                "description": "A GitHub user.",
                "type": "object",
                "properties": {
                    "name": {
                        "nullable": true,
                        "type": "string"
                    },
                    "email": {
                        "nullable": true,
                        "type": "string"
                    },
                    "login": {
                        "type": "string",
                        "example": "octocat"
                    },
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "node_id": {
                        "type": "string",
                        "example": "MDQ6VXNlcjE="
                    },
                    "avatar_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://github.com/images/error/octocat_happy.gif"
                    },
                    "gravatar_id": {
                        "type": "string",
                        "example": "41d064eb2195891e12d0413f63227ea7",
                        "nullable": true
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat"
                    },
                    "html_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://github.com/octocat"
                    },
                    "followers_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/followers"
                    },
                    "following_url": {
                        "type": "string",
                        "example": "https://api.github.com/users/octocat/following{/other_user}"
                    },
                    "gists_url": {
                        "type": "string",
                        "example": "https://api.github.com/users/octocat/gists{/gist_id}"
                    },
                    "starred_url": {
                        "type": "string",
                        "example": "https://api.github.com/users/octocat/starred{/owner}{/repo}"
                    },
                    "subscriptions_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/subscriptions"
                    },
                    "organizations_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/orgs"
                    },
                    "repos_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/repos"
                    },
                    "events_url": {
                        "type": "string",
                        "example": "https://api.github.com/users/octocat/events{/privacy}"
                    },
                    "received_events_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/received_events"
                    },
                    "type": {
                        "type": "string",
                        "example": "User"
                    },
                    "site_admin": {
                        "type": "boolean"
                    },
                    "starred_at": {
                        "type": "string",
                        "example": "\"2020-07-09T00:17:55Z\""
                    }
                },
                "required": [
                    "avatar_url",
                    "events_url",
                    "followers_url",
                    "following_url",
                    "gists_url",
                    "gravatar_id",
                    "html_url",
                    "id",
                    "node_id",
                    "login",
                    "organizations_url",
                    "received_events_url",
                    "repos_url",
                    "site_admin",
                    "starred_url",
                    "subscriptions_url",
                    "type",
                    "url"
                ],
                "nullable": true
            },
            "simple-user": {
                "title": "Simple User",
                "description": "A GitHub user.",
                "type": "object",
                "properties": {
                    "name": {
                        "nullable": true,
                        "type": "string"
                    },
                    "email": {
                        "nullable": true,
                        "type": "string"
                    },
                    "login": {
                        "type": "string",
                        "example": "octocat"
                    },
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "node_id": {
                        "type": "string",
                        "example": "MDQ6VXNlcjE="
                    },
                    "avatar_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://github.com/images/error/octocat_happy.gif"
                    },
                    "gravatar_id": {
                        "type": "string",
                        "example": "41d064eb2195891e12d0413f63227ea7",
                        "nullable": true
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat"
                    },
                    "html_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://github.com/octocat"
                    },
                    "followers_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/followers"
                    },
                    "following_url": {
                        "type": "string",
                        "example": "https://api.github.com/users/octocat/following{/other_user}"
                    },
                    "gists_url": {
                        "type": "string",
                        "example": "https://api.github.com/users/octocat/gists{/gist_id}"
                    },
                    "starred_url": {
                        "type": "string",
                        "example": "https://api.github.com/users/octocat/starred{/owner}{/repo}"
                    },
                    "subscriptions_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/subscriptions"
                    },
                    "organizations_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/orgs"
                    },
                    "repos_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/repos"
                    },
                    "events_url": {
                        "type": "string",
                        "example": "https://api.github.com/users/octocat/events{/privacy}"
                    },
                    "received_events_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/users/octocat/received_events"
                    },
                    "type": {
                        "type": "string",
                        "example": "User"
                    },
                    "site_admin": {
                        "type": "boolean"
                    },
                    "starred_at": {
                        "type": "string",
                        "example": "\"2020-07-09T00:17:55Z\""
                    }
                },
                "required": [
                    "avatar_url",
                    "events_url",
                    "followers_url",
                    "following_url",
                    "gists_url",
                    "gravatar_id",
                    "html_url",
                    "id",
                    "node_id",
                    "login",
                    "organizations_url",
                    "received_events_url",
                    "repos_url",
                    "site_admin",
                    "starred_url",
                    "subscriptions_url",
                    "type",
                    "url"
                ]
            },
            "nullable-milestone": {
                "title": "Milestone",
                "description": "A collection of related issues and pull requests.",
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/repos/octocat/Hello-World/milestones/1"
                    },
                    "html_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://github.com/octocat/Hello-World/milestones/v1.0"
                    },
                    "labels_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.github.com/repos/octocat/Hello-World/milestones/1/labels"
                    },
                    "id": {
                        "type": "integer",
                        "example": 1002604
                    },
                    "node_id": {
                        "type": "string",
                        "example": "MDk6TWlsZXN0b25lMTAwMjYwNA=="
                    },
                    "number": {
                        "description": "The number of the milestone.",
                        "type": "integer",
                        "example": 42
                    },
                    "state": {
                        "description": "The state of the milestone.",
                        "example": "open",
                        "type": "string",
                        "enum": [
                            "open",
                            "closed"
                        ],
                        "default": "open"
                    },
                    "title": {
                        "description": "The title of the milestone.",
                        "example": "v1.0",
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "example": "Tracking milestone for version 1.0",
                        "nullable": true
                    },
                    "creator": {
                        "$ref": "#/components/schemas/nullable-simple-user"
                    },
                    "open_issues": {
                        "type": "integer",
                        "example": 4
                    },
                    "closed_issues": {
                        "type": "integer",
                        "example": 8
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2011-04-10T20:09:31Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2014-03-03T18:58:10Z"
                    },
                    "closed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2013-02-12T13:22:01Z",
                        "nullable": true
                    },
                    "due_on": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2012-10-09T23:39:01Z",
                        "nullable": true
                    }
                },
                "required": [
                    "closed_issues",
                    "creator",
                    "description",
                    "due_on",
                    "closed_at",
                    "id",
                    "node_id",
                    "labels_url",
                    "html_url",
                    "number",
                    "open_issues",
                    "state",
                    "title",
                    "url",
                    "created_at",
                    "updated_at"
                ],
                "nullable": true
            },
...
            "issue": {
                "title": "Issue",
                "description": "Issues are a great way to keep track of tasks, enhancements, and bugs for your projects.",
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "node_id": {
                        "type": "string"
                    },
                    "url": {
                        "description": "URL for the issue",
                        "example": "https://api.github.com/repositories/42/issues/1",
                        "type": "string",
                        "format": "uri"
                    },
                    "repository_url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "labels_url": {
                        "type": "string"
                    },
                    "comments_url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "events_url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "html_url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "number": {
                        "description": "Number uniquely identifying the issue within its repository",
                        "example": 42,
                        "type": "integer"
                    },
                    "state": {
                        "description": "State of the issue; either 'open' or 'closed'",
                        "example": "open",
                        "type": "string"
                    },
                    "state_reason": {
                        "description": "The reason for the current state",
                        "example": "not_planned",
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "completed",
                            "reopened",
                            "not_planned"
                        ]
                    },
                    "title": {
                        "description": "Title of the issue",
                        "example": "Widget creation fails in Safari on OS X 10.8",
                        "type": "string"
                    },
                    "body": {
                        "description": "Contents of the issue",
                        "example": "It looks like the new widget form is broken on Safari. When I try and create the widget, Safari crashes. This is reproducible on 10.8, but not 10.9. Maybe a browser bug?",
                        "type": "string",
                        "nullable": true
                    },
                    "user": {
                        "$ref": "#/components/schemas/nullable-simple-user"
                    },
                    "labels": {
                        "description": "Labels to associate with this issue; pass one or more label names to replace the set of labels on this issue; send an empty array to clear all labels from the issue; note that the labels are silently dropped for users without push access to the repository",
                        "example": [
                            "bug",
                            "registration"
                        ],
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "format": "int64"
                                        },
                                        "node_id": {
                                            "type": "string"
                                        },
                                        "url": {
                                            "type": "string",
                                            "format": "uri"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "color": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "default": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "assignee": {
                        "$ref": "#/components/schemas/nullable-simple-user"
                    },
                    "assignees": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/simple-user"
                        },
                        "nullable": true
                    },
                    "milestone": {
                        "$ref": "#/components/schemas/nullable-milestone"
                    },
                    "locked": {
                        "type": "boolean"
                    },
                    "active_lock_reason": {
                        "type": "string",
                        "nullable": true
                    },
                    "comments": {
                        "type": "integer"
                    },
                    "pull_request": {
                        "type": "object",
                        "properties": {
                            "merged_at": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            },
                            "diff_url": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "html_url": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "patch_url": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "url": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            }
                        },
                        "required": [
                            "diff_url",
                            "html_url",
                            "patch_url",
                            "url"
                        ]
                    },
                    "closed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "draft": {
                        "type": "boolean"
                    },
                    "closed_by": {
                        "$ref": "#/components/schemas/nullable-simple-user"
                    },
                    "body_html": {
                        "type": "string"
                    },
                    "body_text": {
                        "type": "string"
                    },
                    "timeline_url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "repository": {
                        "$ref": "#/components/schemas/repository"
                    },
                    "performed_via_github_app": {
                        "$ref": "#/components/schemas/nullable-integration"
                    },
                    "author_association": {
                        "$ref": "#/components/schemas/author-association"
                    },
                    "reactions": {
                        "$ref": "#/components/schemas/reaction-rollup"
                    }
                },
                "required": [
                    "assignee",
                    "closed_at",
                    "comments",
                    "comments_url",
                    "events_url",
                    "html_url",
                    "id",
                    "node_id",
                    "labels",
                    "labels_url",
                    "milestone",
                    "number",
                    "repository_url",
                    "state",
                    "locked",
                    "title",
                    "url",
                    "user",
                    "author_association",
                    "created_at",
                    "updated_at"
                ]
            },
...
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}